+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2 ()
 
 * Added `rank_member_if` and `rank_member_if_in` methods that allow you to rank a member in the leaderboard based on execution of a function. 
+* Added `rank_members` and `rank_members_in` methods that allow you to rank multiple members in a leaderboard at once by passing in an array of members and scores.
 
 ## 2.1 (2012-11-27)